+++ programs/drift/src/lib.rs
@@ -14,6 +14,7 @@ use crate::state::if_rebalance_config::IfRebalanceConfigParams;
 use crate::state::oracle::PrelaunchOracleParams;
 use crate::state::order_params::{ModifyOrderParams, OrderParams};
 use crate::state::perp_market::{ContractTier, MarketStatus};
+use crate::state::scale_order_params::ScaleOrderParams;
 use crate::state::settle_pnl_mode::SettlePnlMode;
 use crate::state::spot_market::AssetTier;
 use crate::state::spot_market::SpotFulfillmentConfigStatus;
@@ -365,6 +366,13 @@ pub mod drift {
         params: Vec<OrderParams>,
     ) -> Result<()> {
         handle_place_orders(ctx, params)
+    }
+
+    pub fn place_scale_orders<'c: 'info, 'info>(
+        ctx: Context<'_, '_, 'c, 'info, PlaceOrder>,
+        params: ScaleOrderParams,
+    ) -> Result<()> {
+        handle_place_scale_orders(ctx, params)
     }
 
     pub fn begin_swap<'c: 'info, 'info>(