+++ sdk/scripts/grpc-client-test-comparison.ts
@@ -96,9 +96,8 @@ async function initializeGrpcDriftClientV2VersusV1() {
 			token: TOKEN,
 			commitmentLevel: CommitmentLevel.PROCESSED,
 			channelOptions: {
-				'grpc.keepalive_time_ms': 10_000,
-				'grpc.keepalive_timeout_ms': 1_000,
-				'grpc.keepalive_permit_without_calls': 1,
+				grpcKeepAliveTimeout: 1_000,
+				grpcTcpKeepalive: 10_000,
 			},
 		},
 	};